+++ hypervisor/dm/vpci/pci_priv.h
@@ -68,43 +68,12 @@ pci_vdev_write_cfg_u32(struct pci_vdev *vdev, uint32_t offset, uint32_t val)
 	*(uint32_t *)(vdev->cfgdata + offset) = val;
 }
 
-static inline uint32_t pci_vdev_read_cfg(struct pci_vdev *vdev,
-	uint32_t offset, uint32_t bytes)
-{
-	uint32_t val;
+uint32_t pci_vdev_read_cfg(struct pci_vdev *vdev, uint32_t offset, uint32_t bytes);
+void pci_vdev_write_cfg(struct pci_vdev *vdev, uint32_t offset, uint32_t bytes, uint32_t val);
 
-	switch (bytes) {
-	case 1U:
-		val = pci_vdev_read_cfg_u8(vdev, offset);
-		break;
-	case 2U:
-		val = pci_vdev_read_cfg_u16(vdev, offset);
-		break;
-	default:
-		val = pci_vdev_read_cfg_u32(vdev, offset);
-		break;
-	}
-
-	return val;
-}
-
-static inline void pci_vdev_write_cfg(struct pci_vdev *vdev, uint32_t offset,
-	uint32_t bytes, uint32_t val)
-{
-	switch (bytes) {
-	case 1U:
-		pci_vdev_write_cfg_u8(vdev, offset, val);
-		break;
-	case 2U:
-		pci_vdev_write_cfg_u16(vdev, offset, val);
-		break;
-	default:
-		pci_vdev_write_cfg_u32(vdev, offset, val);
-		break;
-	}
-}
-
-void pci_vdev_cfg_handler(struct vpci *vpci, uint32_t in,
-	union pci_bdf vbdf, uint32_t offset, uint32_t bytes, uint32_t *val);
+void pci_vdev_cfg_handler(struct vpci *vpci, uint32_t in, union pci_bdf vbdf, uint32_t offset,
+	uint32_t bytes, uint32_t *val);
+uint32_t pci_pdev_read_cfg(struct pci_pdev *pdev, uint32_t offset, uint32_t bytes);
+void pci_pdev_write_cfg(struct pci_pdev *pdev, uint32_t offset, uint32_t bytes, uint32_t val);
 
 #endif /* PCI_PRIV_H_ */
