+++ hypervisor/dm/vpci/pci_pt.c
@@ -32,79 +32,9 @@
 #include <hypervisor.h>
 #include "pci_priv.h"
 
-
-static spinlock_t pci_device_lock = { .head = 0, .tail = 0 };
-
-
 static inline uint32_t pci_bar_base(uint32_t bar)
 {
 	return bar & PCIM_BAR_MEM_BASE;
-}
-
-static uint32_t pci_pdev_calc_address(union pci_bdf bdf, uint32_t offset)
-{
-	uint32_t addr = (uint32_t)bdf.value;
-
-	addr <<= 8U;
-	addr |= (offset | PCI_CFG_ENABLE);
-	return addr;
-}
-
-static uint32_t pci_pdev_read_cfg(struct pci_pdev *pdev,
-	uint32_t offset, uint32_t bytes)
-{
-	uint32_t addr;
-	uint32_t val;
-
-	spinlock_obtain(&pci_device_lock);
-
-	addr = pci_pdev_calc_address(pdev->bdf, offset);
-
-	/* Write address to ADDRESS register */
-	pio_write(addr, PCI_CONFIG_ADDR, 4U);
-
-	/* Read result from DATA register */
-	switch (bytes) {
-	case 1U:
-		val = pio_read8(PCI_CONFIG_DATA + (offset & 3U));
-		break;
-	case 2U:
-		val = pio_read16(PCI_CONFIG_DATA + (offset & 2U));
-		break;
-	default:
-		val = pio_read32(PCI_CONFIG_DATA);
-		break;
-	}
-	spinlock_release(&pci_device_lock);
-
-	return val;
-}
-
-static void pci_pdev_write_cfg(struct pci_pdev *pdev, uint32_t offset,
-	uint32_t bytes, uint32_t val)
-{
-	uint32_t addr;
-
-	spinlock_obtain(&pci_device_lock);
-
-	addr = pci_pdev_calc_address(pdev->bdf, offset);
-
-	/* Write address to ADDRESS register */
-	pio_write(addr, PCI_CONFIG_ADDR, 4U);
-
-	/* Write value to DATA register */
-	switch (bytes) {
-	case 1U:
-		pio_write8(val, PCI_CONFIG_DATA + (offset & 3U));
-		break;
-	case 2U:
-		pio_write16(val, PCI_CONFIG_DATA + (offset & 2U));
-		break;
-	default:
-		pio_write32(val, PCI_CONFIG_DATA);
-		break;
-	}
-	spinlock_release(&pci_device_lock);
 }
 
 static int vdev_pt_init_validate(struct pci_vdev *vdev)
